refactored VMenu, renamed absolute prop, added doc blocks for some mixins, renamed menu specific mixins
