+++ src/mixins/bootable.js
@@ -1,9 +1,15 @@
+/**
+ * Bootable
+ * @mixin
+ *
+ * Used to add lazy content functionality to components
+ * Looks for change in "isActive" to automatically boot
+ * Otherwise can be set manually
+ */
 export default {
-  data () {
-    return {
-      isBooted: false
-    }
-  },
+  data: () => ({
+    isBooted: false
+  }),
 
   watch: {
     isActive () {
@@ -13,7 +19,11 @@ export default {
 
   methods: {
     showLazyContent (content) {
-      return this.lazy && this.isBooted || !this.lazy ? content : null
+      return this.lazy &&
+        this.isBooted ||
+        !this.lazy
+        ? content
+        : null
     }
   }
 }
